Checkpass version 2. API and test cases update
The API has been updated to check all the constraints together rather than stopping at the failed constraint and returning only the value of the constraint which has failed as a string message. Now all the constraints are evaluated at once and an object is returned which provides the data regarding all the constraints which have failed. The object returned consists of an object against every constraint. This object provides a value which indicates whether the check has passed or failed and also provides a string message which indicates the reason for the failure incase the constraint has failed. This would affect the performance of the API since now all the checks are evaluated for every keystroke but would also provide the necessary flexibilty for consuming the library. The performance slowdown can be mitigated using debouncing