checkpass v2.0.0 | Reverting the source code to version 1 API
The source code is being reverted to the version 1 API because the package had been published without actually updating the lib files. So the package being downloaded and consumed by the users still contains the old API from version 1. The new API will be available in version 3 on github and npm.